# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orchis Foundry API",
    "version": "0.7.0",
    "description": "Self-hosted code platform: git hosting, pull requests, per-user BYO-model supply-chain scanning, and LLM-friendly context endpoints. Authenticate with a personal access token: `Authorization: Bearer orc_pat_<...>`."
  },
  "servers": [{ "url": "/v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "orc_pat" }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/repos/{org}/{name}/commits": {
      "get": { "summary": "List commits", "parameters": [{ "name": "ref", "in": "query", "schema": { "type": "string" } }, { "name": "limit", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "[Commit]" } } },
      "post": { "summary": "Commit file edits from the browser (needs write access)", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "branch": { "type": "string" }, "baseSha": { "type": "string" }, "message": { "type": "string" }, "changes": { "type": "array", "items": { "type": "object", "properties": { "path": { "type": "string" }, "content": { "type": "string" }, "delete": { "type": "boolean" } } } } }, "required": ["branch", "message", "changes"] } } } }, "responses": { "200": { "description": "{sha,branch,created}" } } }
    },
...
    "/repos/{org}/{name}/commits/{sha}": { "get": { "summary": "Commit detail + diff", "parameters": [{ "name": "sha", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "{sha,short,message,author,email,date,additions,deletions,files:[FileDiff]}" } } } },
...
    "/repos/{org}/{name}/chat": { "post": { "summary": "Chat about the repo with your model; pinned files become context", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "chatId": { "type": "integer" }, "pinned": { "type": "array", "items": { "type": "string" } }, "messages": { "type": "array", "items": { "type": "object", "properties": { "role": { "type": "string" }, "content": { "type": "string" } } } } }, "required": ["messages"] } } } }, "responses": { "200": { "description": "{chatId,message}" } } } },
...
    "/repos/{org}/{name}/chats/{id}": { "get": { "summary": "Messages in a chat", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "{id,messages}" } } } },
...
  }
}